doctor schedule, registratio, queue creation
