+++ frontend/src/pages/Admin/DoctorRegistration/DoctorRegistration.scss
