+++ frontend/src/pages/Admin/DoctorRegistration/DoctorRegistration.tsx
